requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Time series analysis
prophet>=1.1.4
statsmodels>=0.14.0

# Async and web
aiohttp>=3.8.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# Metrics and monitoring
prometheus-client>=0.18.0

# Configuration
pyyaml>=6.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database (optional, for metrics storage)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # For PostgreSQL/TimescaleDB

# Utilities
python-dateutil>=2.8.2

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.7.0

